% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\docType{methods}
\name{boxplotContinuous}
\alias{boxplotContinuous}
\title{plot boxplot}
\usage{
boxplotContinuous(
  .data,
  index = NULL,
  color = "default",
  group = "Group",
  signif = F,
  comparlist = NULL
)
}
\arguments{
\item{.data}{data frame}

\item{index}{选择使用的表型}

\item{color}{颜色选择，目前有"aaas","d3"两种}

\item{signif}{是否需要展示差异检验（默认秩和检验）}

\item{comparlist}{秩和检验分组情况}
}
\description{
plot boxplot
}
\examples{

Group age
A-pp  29
A-pp  36
A-pp  22
path = system.file("extdata","metadata.txt", package = "metagenomeR")
comparlist=list(c("A-pp","S-pp"),c("A-sp","S-sp"),c("A-pp","A-sp"),c("S-pp","S-sp"))
data <- read.table(path, header=T, row.names=1, sep="\t", comment.char="",check.names=F, stringsAsFactors = F)
boxplotContinuous(data,index="age",color="aaas", group="Group",signif=T,comparlist=comparlist)
boxplotContinuous(data,index="age", group="Group")

}
\author{
YuDanZhang
}
