% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate alpha-diversity indices from a phyloseq object}
\usage{
calc_alpha(
  ps,
  indices = c("Richness", "Shannon.Effective", "Shannon", "Inverse.Simpson")
)
}
\arguments{
\item{ps}{phyloseq object}

\item{indices}{Alpha-diversity indices to calculate. Supported metrics are: \cr
\code{"Richness"},\code{"Shannon.Effective"}, \code{"Shannon"} and \code{"Inverse.Simpson"}}
}
\value{
Returns specified alpha-diversity metrics as a dataframe
}
\description{
Calculate alpha-diversity indices from a phyloseq object
}
\examples{
data(zeller_2014)
alpha_div <- calc_alpha(zeller_2014)
# Only richness
alpha_div <- calc_alpha(zeller_2014, indices = c("Richness"))

}
