% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{calc_betadiv}
\alias{calc_betadiv}
\title{Calculate Beta Diversity and Ordination on a Phyloseq Object}
\usage{
calc_betadiv(ps, dist, ord_method = "NMDS")
}
\arguments{
\item{ps}{A phyloseq object containing both the feature table and metadata.}

\item{dist}{A character string specifying the distance metric to use. Supported metrics are \code{"bray"} and \code{"jaccard"}.}

\item{ord_method}{A character string specifying the ordination method to use. Supported methods are \code{"NMDS"}, \code{"MDS"}, and \code{"PCoA"}. Default is \code{"NMDS"}.}
}
\value{
A list containing the distance matrix and ordination object.
}
\description{
This function calculates a beta diversity distance matrix and ordination using methods such as NMDS, MDS, or PCoA.
It supports distance metrics such as Bray-Curtis and Jaccard.
}
\note{
If an unsupported ordination method or distance metric is provided, the function will print an error message and stop execution.
}
\examples{
# Example usage:
beta_div_result <- calc_betadiv(ps, dist = "bray", ord_method = "NMDS")

}
