% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{check_equal_sample_sizes}
\alias{check_equal_sample_sizes}
\title{Check if Sample Sizes are Equal Across Groups}
\usage{
check_equal_sample_sizes(df, variable_col)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{variable_col}{A character string representing the column with groupings to check for equal sample sizes.}
}
\value{
A logical value: \code{TRUE} if all groups have equal sample sizes, and \code{FALSE} otherwise.
}
\description{
This function checks whether the sample sizes for each group in a specified column of a data frame are equal.
}
\examples{
# Example usage:
equal_sizes <- check_equal_sample_sizes(df, "group_column")

}
