% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_utils.R
\name{filter_rownames}
\alias{filter_rownames}
\title{Filter Data Frame by Row Names}
\usage{
filter_rownames(df, filt_vector)
}
\arguments{
\item{df}{A data frame to be filtered.}

\item{filt_vector}{A vector of row names to keep in the data frame.}
}
\value{
A data frame filtered to include only the specified row names.
}
\description{
This function filters the rows of a data frame based on a specified vector of row names.
}
\examples{
# Example usage:
filtered_df <- filter_rownames(df, c("row1", "row2"))

}
\keyword{internal}
