% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_summary.R
\name{func_summarise.default}
\alias{func_summarise.default}
\title{Title}
\usage{
\method{func_summarise}{default}(dt, samples, method = "RC", group_by)
}
\arguments{
\item{group_by}{}
}
\description{
Title
}
