% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_summary.R
\name{func_summarise.fst_table}
\alias{func_summarise.fst_table}
\title{func_summarise.fst_table}
\usage{
\method{func_summarise}{fst_table}(dt, samples, method = "RC", group_by)
}
\arguments{
\item{group_by}{}
}
\description{
func_summarise.fst_table
}
