% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_utils.R
\name{get_db_paths}
\alias{get_db_paths}
\title{Get Database Path for Functional Annotations}
\usage{
get_db_paths(db)
}
\arguments{
\item{db}{A character string specifying the type of functional database. Supported options are \code{"KO"}, \code{"EC"}, and \code{"pathways"}.}
}
\value{
A character string representing the path to the functional database file.
}
\description{
This function returns the path to a specific functional annotation database file depending on the operating system and the type of database requested.
}
\examples{
# Example usage:
db_path <- get_db_paths("KO")

}
\keyword{internal}
