% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{get_top_n}
\alias{get_top_n}
\title{Get Top N Taxa by Mean Abundance}
\usage{
get_top_n(ps, n, level = "species", agg = F)
}
\arguments{
\item{ps}{A phyloseq object containing the taxa data.}

\item{n}{An integer specifying the number of top taxa to return.}

\item{level}{A character string specifying the taxonomic level to aggregate taxa. Default is \code{"species"}.}

\item{agg}{A logical value indicating whether to aggregate taxa at the specified level. Default is \code{FALSE}.}
}
\value{
A character vector of the top N taxa by mean abundance.
}
\description{
This function extracts the top N taxa from a phyloseq object based on mean relative abundance.
Optionally, taxa can be aggregated to a specified taxonomic level.
}
\examples{
# Example usage:
top_taxa <- get_top_n(ps, n = 10, level = "species", agg = TRUE)

}
