% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{get_top_n_group}
\alias{get_top_n_group}
\title{Get Top N Taxa by Mean Abundance per Group}
\usage{
get_top_n_group(ps, n, level = "species", var, group = NULL, agg = F)
}
\arguments{
\item{ps}{A phyloseq object containing the taxa data.}

\item{n}{An integer specifying the number of top taxa to return.}

\item{level}{A character string specifying the taxonomic level to aggregate taxa. Default is \code{"species"}.}

\item{var}{A character string specifying the grouping variable in the phyloseq object.}

\item{group}{A character string specifying the group to filter on.}

\item{agg}{A logical value indicating whether to aggregate taxa at the specified level. Default is \code{FALSE}.}
}
\value{
A character vector of the top N taxa by mean abundance for the specified group.
}
\description{
This function extracts the top N taxa from a phyloseq object based on mean relative abundance for a specific group.
Optionally, taxa can be aggregated to a specified taxonomic level.
}
\examples{
# Example usage:
top_taxa_group <- get_top_n_group(ps, n = 10, level = "species", var = "Treatment", group = "Control")

}
