% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_utils.R
\name{import_func_profile}
\alias{import_func_profile}
\title{Import Functional Profile}
\usage{
import_func_profile(
  meta,
  un,
  ec,
  ko,
  py,
  filter_ungrouped = T,
  renorm = T,
  read = F,
  fix_names = T
)
}
\arguments{
\item{meta}{A data frame or file path representing the metadata.}

\item{un}{A data frame or file path representing the UniRef data.}

\item{ec}{A data frame or file path representing the EC (Enzyme Commission) data.}

\item{ko}{A data frame or file path representing the KO (KEGG Orthology) data.}

\item{py}{A data frame or file path representing the pathway data.}

\item{filter_ungrouped}{A logical value indicating whether to filter out ungrouped, unmapped, and unintegrated data. Default is \code{TRUE}.}

\item{renorm}{A logical value indicating whether to renormalize the data after filtering. Default is \code{TRUE}.}

\item{read}{A logical value indicating whether to read the input data from files. Default is \code{FALSE}.}

\item{fix_names}{A logical value indicating whether to clean the sample names by removing suffixes. Default is \code{TRUE}.}
}
\value{
A list containing the processed UniRef, EC, KO, and pathway functional profiles.
}
\description{
This function imports functional profiles (such as UniRef, EC, KO, and pathways), processes them, and applies optional filtering and renormalization.
}
\examples{
# Example usage:
func_profile <- import_func_profile(meta, un, ec, ko, py, read = TRUE)

}
