% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{import_pseq_metag}
\alias{import_pseq_metag}
\title{Import Phyloseq Object from Merged Table and Metadata}
\usage{
import_pseq_metag(merged_table_path, metapath, level, tax_format = "metaphlan")
}
\arguments{
\item{merged_table_path}{A character string representing the file path of the merged count table.}

\item{metapath}{A character string representing the file path of the metadata.}

\item{level}{A character string specifying the taxonomic level to extract (e.g., "Genus", "Species").}

\item{tax_format}{A character string specifying the format of the taxonomic classification. Options are \code{"metaphlan"} or \code{"gtdb"}. Default is \code{"metaphlan"}.}

\item{table_comment_char}{A character string specifying the comment character in the merged table file. Default is \code{"#"}.}

\item{meta_comment_char}{A character string specifying the comment character in the metadata file. Default is \code{"#"}.}
}
\value{
A phyloseq object containing the filtered count data, taxonomic table, and metadata.
}
\description{
This function reads in a merged taxonomic count table and metadata, selects the specified taxonomic level, and converts the data into a phyloseq object. It supports different taxonomic formats such as Metaphlan and GTDB.
}
\examples{
# Example usage:
phyloseq_obj <- import_pseq_metag("path/to/merged_table.txt", "path/to/metadata.txt", level = "Genus")

}
