% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{load_phylo}
\alias{load_phylo}
\title{Load Phyloseq Object from ASV Table, Taxonomy, and Mapping}
\usage{
load_phylo(asvtab, taxa, mapping, tree = NULL)
}
\arguments{
\item{asvtab}{A data frame representing the ASV table.}

\item{taxa}{A data frame representing the taxonomy data.}

\item{mapping}{A data frame representing the sample metadata.}

\item{tree}{A character string specifying the file path of the phylogenetic tree (optional).}
}
\value{
A phyloseq object containing the ASV table, taxonomy, metadata, and optionally the phylogenetic tree.
}
\description{
This function converts ASV (amplicon sequence variant) table, taxonomy, and sample mapping data into a phyloseq object. It also supports adding a phylogenetic tree.
}
\examples{
# Example usage:
phyloseq_obj <- load_phylo(asvtab, taxa, mapping, tree = "tree.nwk")

}
