% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{maaslin2_func}
\alias{maaslin2_func}
\title{Perform MaAsLin2 Analysis for Functional Profiles}
\usage{
maaslin2_func(
  func_profiles,
  feattype,
  out,
  fixed,
  abun_thresh = 0,
  prev_thresh = 0,
  cores = 6,
  ...
)
}
\arguments{
\item{func_profiles}{A list containing functional profiles and corresponding metadata.}

\item{feattype}{A character string specifying the type of feature (e.g., \code{"UNIREF"}, \code{"EC"}, \code{"KO"}, \code{"pathways"}).}

\item{out}{A character string specifying the output directory for MaAsLin2 results.}

\item{fixed}{A character vector of fixed effects (i.e., covariates) to include in the model.}

\item{abun_thresh}{A numeric value specifying the minimum abundance threshold for filtering features. Default is \code{0}.}

\item{prev_thresh}{A numeric value specifying the minimum prevalence threshold for filtering features. Default is \code{0}.}

\item{cores}{An integer specifying the number of cores to use for parallel computation. Default is \code{6}.}

\item{...}{Additional arguments passed to the \code{Maaslin2} function.}
}
\value{
The output is saved to the directory specified by the \code{out} parameter, containing MaAsLin2 results.
}
\description{
This function performs a MaAsLin2 analysis on functional profiles such as UniRef, EC, KO, or pathways.
The function supports multiple feature types and includes filtering options for abundance and prevalence thresholds.
}
\examples{
# Example usage:
maaslin2_func(func_profiles, feattype = "UNIREF", out = "maaslin2_output", fixed = c("age", "sex"))

}
