% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{maaslin2_tax}
\alias{maaslin2_tax}
\title{Perform MaAsLin2 Analysis for Taxonomic Profile}
\usage{
maaslin2_tax(
  tax_profile,
  out,
  fixed,
  abun_thresh = 0,
  prev_thresh = 0,
  taxonomy = "metaphlan",
  ...
)
}
\arguments{
\item{tax_profile}{A phyloseq object or similar containing the taxonomic profiles and metadata.}

\item{out}{A character string specifying the output directory for MaAsLin2 results.}

\item{fixed}{A character vector of fixed effects (i.e., covariates) to include in the model.}

\item{abun_thresh}{A numeric value specifying the minimum abundance threshold for filtering taxa. Default is \code{0}.}

\item{prev_thresh}{A numeric value specifying the minimum prevalence threshold for filtering taxa. Default is \code{0}.}

\item{taxonomy}{A character string specifying the taxonomy database used (e.g., \code{"metaphlan"} or \code{"gtdb"}). Default is \code{"metaphlan"}.}

\item{...}{Additional arguments passed to the \code{Maaslin2} function.}
}
\value{
The output is saved to the directory specified by the \code{out} parameter, containing MaAsLin2 results.
}
\description{
This function performs a MaAsLin2 analysis on a taxonomic profile dataset, allowing for the inclusion
of fixed effects and the specification of abundance and prevalence thresholds.
It supports different taxonomic classifications such as Metaphlan or GTDB.
}
\examples{
# Example usage:
maaslin2_tax(tax_profile, out = "maaslin2_output", fixed = c("age", "sex"))

}
