% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{meta_to_df}
\alias{meta_to_df}
\title{Convert Sample Metadata to Data Frame}
\usage{
meta_to_df(ps, rownames = T)
}
\arguments{
\item{ps}{A phyloseq object containing the sample metadata.}

\item{rownames}{A logical value indicating whether to include row names as a column. Default is \code{TRUE}.}
}
\value{
A data frame containing the sample metadata.
}
\description{
This function converts the sample metadata in a phyloseq object into a data frame.
Optionally, it includes row names as a column.
}
\examples{
# Example usage:
metadata <- meta_to_df(ps, rownames = FALSE)

}
