% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeR-package.R
\docType{package}
\name{metagenomeR-package}
\alias{metagenomeR}
\alias{metagenomeR-package}
\title{metagenomeR: Analysis Tools for Read-Based Shotgun Metagenomic Data}
\description{
Provides a comprehensive toolkit for analyzing read-based shotgun metagenomic data. Includes functions for importing taxonomic and functional profiles from MetaPhlAn and HUMAnN, calculating alpha and beta diversity metrics, performing differential abundance analysis via MaAsLin2, and generating publication-ready visualizations including PCA/PCoA ordination plots, boxplots, scatter plots, and volcano plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamsorbie/metagenomeR}
  \item Report bugs at \url{https://github.com/adamsorbie/metagenomeR/issues}
}

}
\author{
\strong{Maintainer}: Adam Sorbie \email{adamsorbie@pm.me}

}
\keyword{internal}
