% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{metaphlanResult-class}
\alias{metaphlanResult-class}
\title{Class "metaphlanResult"
This class represents the result of metaphlan analysis.}
\description{
Class "metaphlanResult"
This class represents the result of metaphlan analysis.
}
\section{Slots}{

\describe{
\item{\code{phylum}}{门}
}}

\author{
YangWang
}
\keyword{classes}
