% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\docType{methods}
\name{permanova}
\alias{permanova}
\title{permanova}
\usage{
permanova(.data, .metadata, group, groupColName = "Group", parallel = T)
}
\arguments{
\item{.data}{metaphlanResult对象}

\item{.metadata}{.metadata}

\item{group}{group}

\item{groupColName}{groupColName}

\item{parallel}{parallel}
}
\description{
permanova
}
\examples{

mres <- readMetaphlan(system.file("extdata","metaphlan", package = "metagenomeR"))
metadata <- read.table(system.file("extdata","metadata.txt", package = "metagenomeR"), header=T, row.names=1, sep="\t", comment.char="",check.names=F, stringsAsFactors = F)
permanova(mres, metadata,group = c("A-pp","S-pp"),parallel=T)

}
\author{
YangWang
}
