% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\docType{methods}
\name{phenotypeTest}
\alias{phenotypeTest}
\title{Statistical differences in phenotypic grouping were calculated}
\usage{
phenotypeTest(path, continuous = NULL, discrete = NULL)
}
\arguments{
\item{path}{The metadata file path}

\item{continuous}{continuous variable}

\item{discrete}{discrete variable}
}
\description{
Statistical differences in phenotypic grouping were calculated
}
\examples{

pt <- phenotypeTest(system.file("extdata","metadata.txt", package = "metagenomeR"), continuous=c('age'))

}
\author{
YuDanZhang
}
