% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{phyloseq_adonis}
\alias{phyloseq_adonis}
\title{Perform PERMANOVA (Adonis) Analysis on a Phyloseq Object}
\usage{
phyloseq_adonis(ps, dist_matrix, group_variable, ...)
}
\arguments{
\item{ps}{A phyloseq object containing both the feature table and metadata.}

\item{dist_matrix}{A distance matrix calculated using a method like Bray-Curtis or Jaccard.}

\item{group_variable}{A character string specifying the metadata column to be used as the grouping variable.}

\item{...}{Additional arguments to be passed to the \code{adonis2} function.}
}
\value{
A result object from the \code{adonis2} function containing the R-squared and p-values.
}
\description{
This function performs a PERMANOVA (Adonis) test on a phyloseq object using a distance matrix and a grouping variable.
It returns the R-squared and p-values for the relationship between the distance matrix and the grouping variable.
}
\note{
If the metadata contains missing values (NAs) in the \code{group_variable}, the function will print an error message and return \code{NULL}.
}
\examples{
# Example usage:
dist_matrix <- phyloseq::distance(ps, method = "bray")
adonis_result <- phyloseq_adonis(ps, dist_matrix, group_variable = "Treatment")

}
