% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{phyloseq_betadisper}
\alias{phyloseq_betadisper}
\title{Perform Beta Dispersion Analysis on a Phyloseq Object}
\usage{
phyloseq_betadisper(ps, dist_matrix, group_variable, ...)
}
\arguments{
\item{ps}{A phyloseq object containing both the feature table and metadata.}

\item{dist_matrix}{A distance matrix calculated using methods like Bray-Curtis or Jaccard.}

\item{group_variable}{A character string specifying the metadata column to be used as the grouping variable.}

\item{...}{Additional arguments to be passed to the \code{betadisper} function.}
}
\value{
An ANOVA result object from the \code{betadisper} function showing the dispersion among groups.
}
\description{
This function calculates the homogeneity of group dispersions (beta diversity) using a distance matrix and a grouping variable.
It returns the ANOVA results of the dispersion analysis.
}
\note{
If the metadata contains missing values (NAs) in the \code{group_variable}, the function will print an error message and return \code{NULL}.
}
\examples{
# Example usage:
dist_matrix <- phyloseq::distance(ps, method = "bray")
betadisper_result <- phyloseq_betadisper(ps, dist_matrix, group_variable = "Treatment")

}
