% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot Boxplot with Statistical Significance}
\usage{
plot_boxplot(
  df,
  variable_col,
  value_col,
  comparisons_list,
  fill_var = variable_col,
  xlab = variable_col,
  ylab = value_col,
  p_title = NULL,
  multiple_groups = FALSE,
  cols = NULL,
  group.order = NULL,
  paired = FALSE,
  normal = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{variable_col}{A character string representing the independent variable column in the data frame.}

\item{value_col}{A character string representing the dependent variable column in the data frame.}

\item{comparisons_list}{A list of comparisons to be made for significance testing.}

\item{fill_var}{A character string specifying the variable to use for filling the boxplot. Default is the same as \code{variable_col}.}

\item{xlab}{A character string for the x-axis label. Default is the same as \code{variable_col}.}

\item{ylab}{A character string for the y-axis label. Default is the same as \code{value_col}.}

\item{p_title}{A character string for the plot title. Default is \code{NULL}.}

\item{multiple_groups}{A logical value indicating if the plot contains multiple groups. Default is \code{FALSE}.}

\item{cols}{A vector of colors for the boxplot. Default is \code{NULL}.}

\item{group.order}{A character vector specifying the order of the groups. Default is \code{NULL}.}

\item{paired}{A logical value indicating if the test is paired. Default is \code{FALSE}.}

\item{normal}{A logical value indicating if normality should be assumed. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to the plotting functions.}
}
\value{
A ggplot object representing the boxplot with optional statistical significance annotations.
}
\description{
This function generates a boxplot for a given data frame and performs statistical tests to visualize significance. It supports plotting for paired or unpaired tests with options for multiple groups.
}
\examples{
# Example usage:
plot <- plot_boxplot(df, "group", "value", comparisons_list = list(c("A", "B")))

}
