% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA (Principal Component Analysis) for Functional Profile}
\usage{
plot_pca(func_profile, what, var, colours, add_ellipse = F)
}
\arguments{
\item{func_profile}{A list containing functional profiles and corresponding metadata.}

\item{what}{A character string specifying which part of the functional profile to use for PCA.}

\item{var}{A character string specifying the grouping variable in the metadata.}

\item{colours}{A vector of colors to use for the groups.}

\item{add_ellipse}{A logical value indicating whether to add ellipses around the groups. Default is \code{FALSE}.}
}
\value{
A ggplot object representing the PCA plot.
}
\description{
This function calculates and plots a PCA from a functional profile with metadata, optionally adding ellipses for grouping.
}
\examples{
# Example usage:
pca_plot <- plot_pca(func_profile, what = "UNIREF", var = "Condition", colours = c("red", "blue"), add_ellipse = TRUE)

}
