% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{plot_pcoa}
\alias{plot_pcoa}
\title{Plot PCoA (Principal Coordinates Analysis) or NMDS (Non-metric Multidimensional Scaling)}
\usage{
plot_pcoa(
  func_profile,
  what,
  var,
  colours,
  add_ellipse = F,
  size = 3,
  plot = "MDS"
)
}
\arguments{
\item{func_profile}{A list containing functional profiles and corresponding metadata.}

\item{what}{A character string specifying which part of the functional profile to use for PCoA or NMDS.}

\item{var}{A character string specifying the grouping variable in the metadata.}

\item{colours}{A vector of colors to use for the groups.}

\item{add_ellipse}{A logical value indicating whether to add ellipses around the groups. Default is \code{FALSE}.}

\item{size}{A numeric value specifying the size of the points. Default is \code{3}.}

\item{plot}{A character string specifying the type of plot, either \code{"MDS"} (default) or \code{"NMDS"}.}
}
\value{
A ggplot object representing the PCoA or NMDS plot.
}
\description{
This function performs and plots PCoA or NMDS from a functional profile with metadata. It supports adding ellipses and specifying plot dimensions.
}
\examples{
# Example usage:
pcoa_plot <- plot_pcoa(func_profile, what = "KO", var = "Treatment", colours = c("red", "green"), add_ellipse = TRUE)

}
