% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Scatter Plot with Optional Correlation}
\usage{
plot_scatter(
  df,
  x,
  y,
  point_color,
  line_color,
  fill_color,
  xlab,
  ylab,
  corr.method = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{x}{A character string representing the column to be used for the x-axis.}

\item{y}{A character string representing the column to be used for the y-axis.}

\item{point_color}{A character string specifying the color of the points.}

\item{line_color}{A character string specifying the color of the regression line.}

\item{fill_color}{A character string specifying the fill color of the regression line.}

\item{xlab}{A character string for the x-axis label.}

\item{ylab}{A character string for the y-axis label.}

\item{corr.method}{A character string specifying the correlation method (e.g., \code{"pearson"}, \code{"spearman"}, \code{"kendall"}). Default is \code{NULL}.}

\item{...}{Additional arguments passed to the \code{stat_cor} function for correlation statistics.}
}
\value{
A ggplot object representing the scatter plot, with optional correlation results.
}
\description{
This function generates a scatter plot with an optional linear regression line and calculates correlation statistics if specified.
}
\examples{
# Example usage:
plot <- plot_scatter(df, x = "height", y = "weight", point_color = "blue", line_color = "red", fill_color = "lightblue", xlab = "Height", ylab = "Weight", corr.method = "pearson")

}
