% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{plot_taxonomic_comp}
\alias{plot_taxonomic_comp}
\title{Plot Taxonomic Composition as Stacked Bar Chart}
\usage{
plot_taxonomic_comp(
  ps,
  tax_level,
  var,
  ord = NULL,
  n_taxa = 10,
  per_group = F,
  groups = NULL,
  agg = F
)
}
\arguments{
\item{ps}{A phyloseq object containing the taxonomic data.}

\item{tax_level}{A character string specifying the taxonomic rank to plot.
Must be one of: "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species".}

\item{var}{A character string or symbol specifying the metadata variable to facet by.}

\item{ord}{A character vector specifying the order of factor levels for the grouping
variable. Default is \code{NULL} (uses default ordering).}

\item{n_taxa}{An integer specifying the number of top taxa to display. Default is \code{10}.}

\item{per_group}{A logical value indicating whether to select top taxa per group
rather than overall. Default is \code{FALSE}.}

\item{groups}{A character vector of group names to use when \code{per_group = TRUE}.
Default is \code{NULL}.}

\item{agg}{A logical value indicating whether to aggregate taxa at the specified
taxonomic level before plotting. Default is \code{FALSE}.}
}
\value{
A ggplot object representing the stacked bar chart of taxonomic composition.
}
\description{
This function creates a stacked bar chart showing the relative abundance of taxa
at a specified taxonomic level, faceted by a grouping variable.
}
\examples{
# Example usage:
# comp_plot <- plot_taxonomic_comp(ps, "Phylum", "Treatment", n_taxa = 8)
# comp_plot <- plot_taxonomic_comp(ps, "Genus", "Group", ord = c("Control", "Treatment"))

}
