% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Volcano Plot for Differential Analysis}
\usage{
plot_volcano(results_df, pthresh = 0.05, FCthresh = 0.58, ...)
}
\arguments{
\item{results_df}{A list containing the results of a differential analysis, including log2 fold changes and adjusted p-values.}

\item{pthresh}{A numeric value specifying the p-value cutoff for significance. Default is \code{0.05}.}

\item{FCthresh}{A numeric value specifying the log2 fold change cutoff for significance. Default is \code{0.58}.}

\item{...}{Additional arguments passed to the \code{EnhancedVolcano} function.}
}
\value{
A volcano plot highlighting significant features based on fold change and p-value thresholds.
}
\description{
This function generates a volcano plot from a results data frame, showing fold changes and adjusted p-values. It highlights significant features based on specified thresholds.
}
\examples{
# Example usage:
volcano_plot <- plot_volcano(results_df, pthresh = 0.05, FCthresh = 0.58)

}
