% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{ps_to_feattab}
\alias{ps_to_feattab}
\title{Convert OTU Table to Data Frame}
\usage{
ps_to_feattab(ps)
}
\arguments{
\item{ps}{A phyloseq object containing the OTU table.}
}
\value{
A data frame representing the OTU table.
}
\description{
This function converts the OTU (operational taxonomic unit) table of a phyloseq object into a data frame.
}
\examples{
# Example usage:
otu_table <- ps_to_feattab(ps)

}
