% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_utils.R
\name{read_tab_delim_metag}
\alias{read_tab_delim_metag}
\title{Read a Tab-Delimited Metagenomics File}
\usage{
read_tab_delim_metag(df, comment_char)
}
\arguments{
\item{df}{A character string representing the file path of the tab-delimited file.}

\item{comment_char}{A character string specifying the comment character to use. Default is \code{"#"}.}
}
\value{
A data frame containing the data from the tab-delimited file.
}
\description{
This function reads a tab-delimited metagenomics file into a data frame using specific parameters such as row names and custom comment characters.
}
\examples{
# Example usage:
metag_data <- read_tab_delim_metag("data/tab_delimited_file.txt")

}
