% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\docType{methods}
\name{readMetaphlan}
\alias{readMetaphlan}
\title{readMetaphlan}
\usage{
readMetaphlan(path, checkColnames = T)
}
\arguments{
\item{path}{metaphlan输出的文件夹}

\item{checkColnames}{检测样本名称是否相同}
}
\description{
readMetaphlan
}
\examples{

path = system.file("extdata","metaphlan", package = "metagenomeR")
mres <- readMetaphlan(path)

}
\author{
YangWang
}
