% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_utils.R
\docType{data}
\name{stacked_bar_palette}
\alias{stacked_bar_palette}
\title{Custom colour palette for stacked barplots}
\format{
A character vector of length 30 containing hex colour codes.
}
\usage{
stacked_bar_palette
}
\description{
A character vector of 30 hex colour codes designed for visualizing
taxonomic composition in stacked barplots. The first 10 colours are
from the NPG (Nature Publishing Group) palette, with additional
distinguishable colours for datasets with many taxa.
}
\examples{
# Use with ggplot2 scale_fill_manual
# scale_fill_manual(values = stacked_bar_palette)

# Preview the palette
scales::show_col(stacked_bar_palette)

}
\keyword{datasets}
