% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{stat_plot}
\alias{stat_plot}
\title{Perform Statistical Tests and create dataframe to add to plot results}
\usage{
stat_plot(
  df,
  formula,
  variable_col,
  assume_normality = F,
  multiple_groups = F,
  paired = F
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{formula}{A formula specifying the relationship between dependent and independent variables.}

\item{variable_col}{A character string representing the column with groupings.}

\item{assume_normality}{A logical value indicating whether to assume normality for parametric tests. Default is \code{FALSE}.}

\item{multiple_groups}{A logical value indicating whether multiple groups are present. Default is \code{FALSE}.}

\item{paired}{A logical value indicating whether the test is paired. Default is \code{FALSE}.}
}
\value{
A data frame with the statistical test results, including p-values and significance.
}
\description{
This function performs statistical tests (such as Wilcoxon, Kruskal-Wallis, or Friedman) on a data frame based on the provided formula. It also checks for paired tests and multiple groups.
}
\note{
The function will stop if the sample sizes for paired tests are not equal.
}
\examples{
# Example usage:
stat_results <- stat_plot(df, formula = y ~ x, variable_col = "group")

}
