% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_utils.R
\name{t_df}
\alias{t_df}
\title{Transpose Data Frame}
\usage{
t_df(x)
}
\arguments{
\item{x}{A data frame to be transposed.}
}
\value{
A transposed version of the input data frame.
}
\description{
This function transposes a data frame, converting rows to columns and vice versa.
}
\examples{
# Example usage:
transposed_df <- t_df(original_df)

}
\keyword{internal}
