% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_summary.R
\name{taxa_summarise.fst_table}
\alias{taxa_summarise.fst_table}
\title{taxa_summarise.fst_table}
\usage{
\method{taxa_summarise}{fst_table}(dt, samples, method = "RC", group_by = "phylum")
}
\arguments{
\item{group_by}{}
}
\description{
taxa_summarise.fst_table
}
