% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_utils.R
\name{taxonomy}
\alias{taxonomy}
\title{Extract Taxonomy Table from a Phyloseq Object}
\usage{
taxonomy(ps)
}
\arguments{
\item{ps}{A phyloseq object containing the taxonomy data.}
}
\value{
A data frame representing the taxonomy table.
}
\description{
This function extracts the taxonomy table from a phyloseq object and converts it into a data frame.
}
\examples{
# Example usage:
data(zeller2014)
taxa <- taxonomy(zeller2014)

}
