% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_filtering.R
\name{transform_tax}
\alias{transform_tax}
\title{Transform taxonomic abundance data}
\usage{
transform_tax(ps, transform = "relative", offset = 1)
}
\arguments{
\item{ps}{A phyloseq object containing taxonomic abundance data.}

\item{transform}{Character string specifying the transformation to apply.
One of "relative", "arcsin", or "log10". Default is "relative".}

\item{offset}{Numeric offset to add before log10 transformation to avoid
log of zero. Default is 1.}
}
\value{
A phyloseq object with transformed abundance data.
}
\description{
Apply transformations to taxonomic abundance data in a phyloseq object.
}
