% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_stats.R
\name{xyform}
\alias{xyform}
\title{Create a Formula for Linear Models}
\usage{
xyform(y_var, x_vars)
}
\arguments{
\item{y_var}{A character string representing the dependent variable.}

\item{x_vars}{A character vector of independent variables.}
}
\value{
A formula object of the form \code{y_var ~ x_vars}.
}
\description{
This function generates a formula to be used in linear models or statistical tests. It constructs the formula in the format \code{y ~ x1 + x2 + ...}.
}
\keyword{internal}
