% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeller2014}
\alias{zeller2014}
\title{Example metagenomic data}
\format{
An object of class \code{"phyloseq"}.
}
\usage{
data("zeller2014")
}
\description{
A metagenomic dataset of CRC patients and controls from Zeller et al 2014
}
\examples{
\dontrun{
library(metagenomeR)
data("zeller2014")
print(zeller2014)
}

}
\references{
\itemize{
\item{}{Zeller, G., Tap, J., Voigt, A. Y., Sunagawa, S., Kultima, J. R.,
Costea, P. I., Amiot, A., Böhm, J., Brunetti, F., Habermann, N., Hercog, R.,
Koch, M., Luciani, A., Mende, D. R., Schneider, M. A., Schrotz‐King, P.,
Tournigand, C., Van Nhieu, J. T., Yamada, T., . . . Bork, P. (2014).
Potential of fecal microbiota for early‐stage detection of colorectal cancer.
Molecular Systems Biology, 10(11), 766
\url{https://link.springer.com/article/10.15252/msb.20145645}}
}
}
\keyword{datasets}
