% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate alpha-diversity indices from a phyloseq object}
\usage{
calc_alpha(ps, ...)
}
\arguments{
\item{ps}{phyloseq object}
}
\value{
Returns specified alpha-diversity metrics as a dataframe
}
\description{
Calculate alpha-diversity indices from a phyloseq object
}
\examples{
data(zeller2014)
alpha_div <- calc_alpha(zeller2014)
# Only richness
alpha_div <- calc_alpha(zeller2014, indices = c("Richness"))

}
